Adds Code cell component
